Move context manager into file with expansion functions.
